+++ src/preview/hover.js
@@ -1,0 +1,14 @@
+const vscode = require('vscode');
+
+module.exports = pair => {
+  let contentText =
+`${'```'}
+${pair.openingLineIndex + 1}: ${pair.openingLineText.trim()}
+${'```'}`; // Keep formatting
+
+  const hover = new vscode.Hover(new vscode.MarkdownString(contentText));
+
+  return {
+    hover
+  };
+}