+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release Notes
 
+## [1.2.0]
+Customization Options
+- Preview in line
+- Preview as hover
+- Preview always
+- Preview color
+- Preview italic
+- Preview trigger
+- Disable logging
+
+Fixes 
+- Closing tag not always detected correctly
 ## [1.1.2]
 - Support code format, where open curly brackets are put on a new line => [Preview multiple lines? #4](https://github.com/j0meinaster/bracket-peek/issues/4)
 ## [1.1.1]